File Changes
In the should_return_not_found_when_update_passport_by_non_existent_id
method, I corrected the value of the Long variable, now such a value can
hardly be in the database.
In the should_update_passport_by_id method, I removed the variable with
the Long type and now, before changing the entity, I first save it to
the database using a test in which I create and save the entity.
Also fixed a conflict with the master main branch