Initial commit
Added all files. About the changes: I made idempotence at the controller
level. In short, with several post requests (transmitting information to
create an entity), an error is returned from the 429 controller (Too
many requests), i.e. if such a sequence of data is transmitted to the
server for the first time, the server saves the entity with such data
and returns the Created response, and if an entity with such data
exists, the server returns Too many requests. I also prescribed new
tests for the controller. An important feature in them is that the id
field, when deleting an entity in one test, is auto-incremented in
subsequent tests (+1). Example: I am in the test of creating an entity,
first I create it, then I delete it, everything passes. In the next
test, when updating an entity, I first try to create it with the same id
and already throws an error, because after deleting an entity, for
example, with id = 4, the next entity has id = 5 (4+1), apparently this
is a feature of the database.