Initial commit
I changed the model (entity) and added a new status field there. I also
updated the Rest controller and Service for the status field. I also
added an sql file and made changes to the database using liquibase. In
addition to all this, I also rewrote the tests for the modified files.